"Change package names from 'oreilly' to 'kousenit' across all modules"
The package names for all the modules have been updated from 'oreilly' to 'kousenit'. This change is necessary to align with the new naming convention standard for our project. It doesn't affect any functionalities within the application, only the references to the packages have been updated accordingly in all parts of the codebase.