Add setUp methods for improved testing reliability
Added setUp methods in test classes to perform a HEAD request on the related API and ensure it's available before running tests, increasing the reliability of the tests. Also adjusted the HTTP status checking in JokeServiceTest to use constant instead of a magic number for readability, and improved exception handling when checking API availability in the GeocoderServiceTest.