Handle WebClientRequestExceptions in service tests
Updated JsonPlaceholderServiceTest, JsonPlaceholderTest and AstroServiceTest to handle WebClientRequestExceptions thrown when the external service is not available. Prior to sending GET requests in tests, a HEAD request is now made to make sure the service used in tests is accessible. If the service is not accessible, tests are skipped using JUnit5's assumeTrue. This prevents tests from failing when the problem is with the external service and not the code being tested.