Refactor Testing setup and Improve HTTP request handling
Refactoring was done in setup functions of the test files by streamlining the HEAD requests to check the availability of APIs. HTTPClient methods are used to write less and cleaner code.

In MyProperties config file, a debug print statement was removed for cleaner log prints.

There are changes to replace WebTestClient with more suitable classes for HTTP request handling for test scenarios. Some assertions were added to check the response of HEAD requests in JsonPlaceholderServiceTest.

Minor changes in GeocoderServiceTest to check the headers received from a HEAD request.

Reason for changes was to improve readability and efficiency of the test code and ensure the robustness of HTTP requests.