Update HTTP request method in JsonPlaceholderService
Changed the HTTP request method from GET to HEAD in JsonPlaceholderService. Removed the @GetExchange() annotation and replaced with @HttpExchange(method = "HEAD"). This ensures that a HEAD request is made instead of a GET request, as the previous attempt to switch to a HEAD request with ResponseEntity<Void> did not work as expected.