Refactor AstroService and AstroServiceTest for better readability
The commit revises the variable names in AstroService to be more descriptive and thus improving the readability of the code. WebClient "client" has been refactored to "webClient" and RestTemplate "template" to "restTemplate". The construction of the restClient object has also been modified to use the factory method `create()`, making it clearer.

In AstroServiceTest, assertion methods have been grouped using assertAll() to logically group related assertions. This allows tests to continue even if one fails, offering a complete report of all failing assertions.