Add RestClient and refactor tests
Add RestClient as another way to make HTTP client calls and refactor code accordingly in AstroService.java. Renamed template to restTemplate and client to webClient to make the names more meaningful. Added configuration for RestClient dependency in build.gradle. Updated AstroServiceTest to include assertions for the new RestClient call, and refactored assertions in the existing tests to use the assertAll method. The AppConfig class is also modified to adjust to the WebClient name change. These changes are done to improve code readability and provide another way to do HTTP client calls.