Add suppression for deprecated removal warning
A "@SuppressWarnings("removal")" annotation was added to the jsonPlaceholderService method in AppConfig.java. This change suppresses compiler warnings related to the use of deprecated items. It was deemed necessary as the deprecated method currently used will not immediately affect the functionality of the application. We can replace it in the future refactorings.