Update Spring Boot version and integrate Astro API
Spring Boot's version was updated from 3.2.0 to 3.2.3, and Gradle version from 8.4 to 8.5. Most notably, Astro API was integrated in the application through the creation of the AstroInterface and its corresponding test class. This new feature enables the retrieval of astrological data, and it has been properly tested to ensure correct implementation.