Update Spring Boot version, refactor tests and add parallel execution option
We upgraded the Spring Boot version from 3.2.0 to 3.2.3 in our build.gradle file. Additionally, we performed some refactoring in the JsonPlaceholderServiceTest suite to improve readability and maintainability. We also added the option for parallel test execution by adding it to our junit-platform.properties file, still in an experimental phase for now.