Update test lifecycle and enable parallel execution
Multiple changes were made to improve test suite execution. The lifecycle of the test class was updated to "per method" for improved test isolation, which includes the creation of a new test instance for each test method. Additionally, integrated support for parallel test execution was activated in the JUnit platform properties, which would aid in reducing total test run time. Also updated the Spring Boot version and the Java source compatibility in the build.gradle file.